Error Banner color fix
